+++ react/components/commons/SkeletonLoader.tsx
@@ -1,8 +1,8 @@
 import React, { FunctionComponent } from 'react'
 
 interface Props {
-  width?: string
-  size?: string
+  width?: '10' | '20' | '25' | '30' | '33' | '34' | '40' | '50' | '60' | '70' | '75' | '80' | '90' | '100' | 'third' | 'two-thirds'
+  size?: '0' | '1' | '2' | '3' | '4' | '5' | '6'
 }
 
 const SkeletonLoader: FunctionComponent<Props> = ({ width = '100', size }) => (